Run `npm run lint:fix` to get all files to the same standard.
We have not been enforcing the built-in prettier standards across the team, which has lead some files to be out of date formatting wise. Running the linter globally makes sure we have the codebase up to the relevant standard.

To make sure we don't fail again in the future a GitHub Action is also added. It will run `npm run lint` after every pull request in the future marking the commit with a red `X` if it fails.