Standardise API design - URLs
We had used a wide range of api endpoints during development, some `camelCased`, some `snake_cased`, while also being liberal on the naming of the endpoints. This PR changes the naming to be more closer to the REST standards using the `/<resource>/<action_or_id>` nomenclature, where resource is always in plural, and the `action_or_id` part either refers to what kind of special function the endpoint does - or what kind of resource needs to be obtained (e.g. a `random` one)