Fix export button functionality
The old export button was done in a very complex way. We first downloaded the data from the URL into the browser's memory then asked the browser to save that file. We could have just asked the browser to download the URL to a file, which as it uses no Javascript is much more performant