Frontend component cleanup - props and state changes
First part of cleaning up the react components. The main issue we had that we had too many props, with a lot of them doing the same or very similar things. This led us to have features broken, as in some cases only part of the props that were needed have been updated downstream in another component.

To solve this issue a couple changes have been introduced:

- Set default values to be "emptier". We used strings like "Loading..." as default values instead of `null` or `false` or similar. This has been changed to make sure they are all an empty-like value.
- Merge `sentence` and `sentenceId`. We don't need to use primitives, we can use larger objects in the props that contain the entire object - both the text and the ID
- Merge `selectedSuggestion` and `enableDisable`. We can only enable the submit button if there's a selected suggestion, so the easy way out is to simply check if we have selected a sugestion, so we don't need to keep track to enable / disable the button separately
- Merge `suggestions` and `loading`. We know that the suggestions are loading if they don't exist yet.
- Loading up a new sentence is a fairly complex bit with multiple states needing to change. This has therefore been extracted into it's own functon and made sure that this is the function that gets called when we need a new sentence, instead of updating all of the state props in every place that needs this functionality.
- Clean up some naming conventions along the line as well, we used `sentence`, `randomText`, `text` for the same value for example throughout the components.

Part of this cleanup some components related to the API have changed - made sure that the API uses `snake_case` conventions, and a bit nicer design for the `/api/user_interactions` call