Frontend Component Cleanup - Reorganize the Alert box
We had a nice popup alert box that showed if the upload was successful or not. Unfortunately it was too deep down in the component list, and was immediately removed as a component when reloading the next sentence. This has been refactored to move that component out and allowing the callers to set the alert message as they like