CSS updates - Responsive Design
Responsive design is making sure that the website looks accessible on all kinds of screen sizes from large monitors down to small mobile screens. The easiest way to test this is to simply start resizing your browser window - if the app looks nice regardless of how you resize the window then it is considered responsive.

An easy way to add responsiveness is through media queries in the CSS, making sure to only update details that don't look nice. In this example we decreased the margin once we started decreasing the window size, and then switched to single column grid layout from the dual-column one