CSS updates - Cleanup
Going through all of the CSS classes removing any that are not used anymore. Also making sure to remove classes that are enforcing their own name (e.g. `paddingBottom` adds padding to the bottom). Unless the design is for a framework it's preferred to use css classes that describe what the object is and not how it looks like. This allows people to update the "how" wihtout causing canfusion. Imagine if due to a change you need to switch `.fullWidth` to only mean `width: 80%` and not `width: 100%`.