Frontend Compoment Cleanup - Popups again
Moving the popups one layer above so they are available throughout the pages. Also make sure we can create a popup from the URL bar. This is important as the file upload feature doesn't use JavaScript so we need to roll back to plain HTTP/HTML to be able to communicate success and failure from the upload