Improve Admin features - show statistics
Final addition to the admin pages - show statistics of the selected sentence. We added a couple endpoints to the API that allows us to return some statistics about the sentence, including how many people clicked on the various suggestions and other interactions.

Indices have been added to make sure the databaase is performance enough even if we have a lot of suggestions