Remove duplications from test
We have introduced a lot of duplication in the tests. This makes the automated tests more brittle - a single change in the codebase might mean you need to change a lot of tests to match the change. Having more deduplicated tests help show the user the common features as well making it easier to see what is the same and what is different.