Add preliminary unit testing for the frontend
Similar to the backend unit tests we want to make sure we can test frontend component as well. We need to split the two into separate Jest projects as the configuration on how they work will be different. Once the split we add some simple unit tests to the `Navbar` component to make sure the testing framework works.