Add end-to-end tests using the Cucumber BDD framework
Cucumber is a BDD - Behavioural Driven Development framework that allows feature files written in plain English to be used as test cases, check the `sending_suggestions.feature` for an example how it looks like.

Cucumber and generally BDD frameworks are commonyly used to describe end-to-end tests, these tests are used on a running instance and checks the interactions between the entire stack: browser, frontend, backend and database.

This change does the following:

* Adds general support for the cucumber framework
* Add an initial set of tests covering the core functionality of the page - submitting suggestions
* Makes sure GitHub will run the tests at each PR to make sure features are still working as expected
* Updates the README around testing