Final set of E2E features - testing login and admin features
For login tests we use the same trick we've done with unit tests - login is done using a test-only login endpoint, that allows us to mock any user. This avoids the tests to be tied with Google, although this also means we don't have proper E2E tests actually checking Google login.

For admin tests we needed to do a couple tricks for the download feature - we redirected the browser to download files in a known directory then we simply check if the file appeared over there or not.